requirements.txt
asgiref==3.5.0
boto3==1.21.32
botocore==1.24.32
dj-database-url==0.5.0
Django==3.2
django-allauth==0.41.0
django-countries==7.3.2
django-crispy-forms==1.14.0
django-storages==1.12.3
gunicorn==20.1.0
jmespath==1.0.0
oauthlib==3.2.0
Pillow==9.0.1
psycopg2-binary==2.9.3
python3-openid==3.2.0
pytz==2021.3
requests-oauthlib==1.3.1
s3transfer==0.5.2
sqlparse==0.4.2
stripe==2.68.0
